apps/backend/pyproject.toml
[tool.poetry]
name = "landbiznes-backend"
version = "1.0.0"
description = "Enterprise-grade FastAPI backend for LandBiznes"
authors = ["LandBiznes Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
aioredis = "^2.0.1"
geoalchemy2 = "^0.14.0"
shapely = "^2.0.0"
pydantic-extra-types = "^2.0.0"
email-validator = "^2.1.0"
aiosqlite = "^0.19.0"
asyncpg = "^0.29.0"
httpx = "^0.25.0"
tenacity = "^8.2.0"
sqlalchemy-utils = "^0.41.0"
pytz = "^2023.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.build"
